requirements.txt
python-dotenv
pandas
lxml
argparse
tqdm
PyYAML
requests
